Fixed Japanese Vernal and Autumnal Equinox days  (#116)
* Fixed Japanese Equinox Day calculation

There was an algorithm to calculate Vernal and Autumnal Equinox Days for Japanese calendar. Although specific days for these holidays are announced every year by the government, it has never announced a different date.

* Fixed Japanese Vernal and Autumnal Equinox days

Fixed case when Vernal or Autumnal Equinox day falls on Sunday. In this case the next Monday becomes a holiday.

---------

Co-authored-by: Mazov Sergey <mazov@indexpro.co.jp>