New MexicoPublicHoliday class with Mexican holidays. (#158)
New GetDayOfWeekInMonth method of the HolidayCalculator class.

Since Mexico's presidential inauguration only happens every 6 years and is a holiday, this method performs the calculation.

Inclusion of the acronym "Mx" in the PublicHolidayCountryCode class.

Inclusion of MexicoPublicHoliday in the GetPublicHolidayForCountry method of the PublicHolidayFactory class.

New TestMexicoPublicHoliday test class.

Inclusion of Mexico in the tests of the TestAllPublicHolidays class.